backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# AI Services
anthropic==0.7.8
openai==1.3.7

# Email
fastapi-mail==1.4.1
jinja2==3.1.2

# File Handling
python-magic==0.4.27
pillow==10.1.0

# Validation & Serialization
pydantic==2.5.0
email-validator==2.1.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Development
watchfiles==0.21.0
python-jose==3.3.0
